Revert "Second try (#74)"
This reverts commit 7b8e32ece8964809a022d4069a734eb16cf1126b.
